code reorg. Toolbox.Render separated out.
